# Conflicts:
#	doyto-query-jdbc/src/main/java/win/doyto/query/jdbc/DatabaseTemplate.java
#	doyto-query-jdbc/src/main/java/win/doyto/query/jdbc/JdbcAssociationService.java
#	doyto-query-jdbc/src/main/java/win/doyto/query/jdbc/JdbcDataAccess.java

+++ doyto-query-test/src/main/java/win/doyto/query/test/menu/MenuView.java
@@ -32,10 +32,10 @@ import java.util.List;
  */
 @Getter
 @Setter
-public class MenuView implements Persistable<Long> {
+public class MenuView implements Persistable<Integer> {
 
     @Id
-    private Long id;
+    private Integer id;
     private String menuName;
     private String platform;
